Add bubbling core-signal events in addition to element based events
In order to make supporting elements, access to the app-router events is needed. Sending core-signal events allows elements that are ignorant of the rest of the application still respond meaningfully to these events.

This is needed for elements such as page tracking, comment plug-ins where the provider requires a singleton call, or other sitewide changes based on route changes.